LPD-79533 Improve design library settings code and change styles to match designs
