+++ modules/apps/design-library/design-library-web/src/main/resources/META-INF/resources/js/props_transformer/cell_renderers/createSetItemComponentProps.tsx
@@ -5,14 +5,14 @@
 
 import {dateUtils} from 'frontend-js-web';
 
-import {ActionItem, DesignLibraryItem} from '../../types';
+import {ActionItem, DesignLibrary} from '../../types';
 
 export default function createSetItemComponentProps(symbol: string) {
 	return function ({
 		item,
 		props,
 	}: {
-		item: DesignLibraryItem;
+		item: DesignLibrary;
 		props: {actions: ActionItem[]};
 	}) {
 		const editAction = props.actions.find((action) => {
