+++ modules/apps/design-library/design-library-web/src/main/resources/META-INF/resources/js/types.ts
@@ -8,44 +8,25 @@ export interface ActionItem {
 	href?: string;
 }
 
-export interface DesignLibraryItem {
-	creatorUserId: number;
-	dateModified: string;
-	description: string;
+interface Creator {
+	additionalName: string;
+	contentType: string;
 	externalReferenceCode: string;
+	familyName: string;
+	givenName: string;
 	id: number;
 	name: string;
 }
 
 export interface DesignLibrary {
 	assetLibraryKey: string;
-	creatorUserId: string;
+	creator: Creator;
+	dateModified: string;
 	description: string;
 	externalReferenceCode: string;
 	id: number;
 	name: string;
-	settings?: DesignLibrarySettings;
 	siteId: number;
-}
-
-export interface DesignLibrarySettings {
-	availableLanguageIds?: string[];
-	defaultLanguageId?: string;
-	logoColor?:
-		| 'outline-0'
-		| 'outline-1'
-		| 'outline-2'
-		| 'outline-3'
-		| 'outline-4'
-		| 'outline-5'
-		| 'outline-6'
-		| 'outline-7'
-		| 'outline-8'
-		| 'outline-9';
-	sharingEnabled?: boolean;
-	trashEnabled?: boolean;
-	trashEntriesMaxAge?: number;
-	useCustomLanguages?: boolean;
 }
 
 export interface Site {
