+++ modules/apps/design-library/design-library-web/src/main/resources/META-INF/resources/view_settings.jsp
@@ -17,5 +17,5 @@ designLibrarySettingsDisplayContext.setPortletDisplay(portletDisplay, renderResp
 
 <react:component
 	module="{DesignLibrarySettings} from design-library-web"
-	props="<%= designLibrarySettingsDisplayContext.getSettingsProps() %>"
+	props="<%= designLibrarySettingsDisplayContext.getProps() %>"
 />
